Big update

+++ src/models/category.model.js
@@ -1,22 +1,28 @@
 import { v7 as uuidv7 } from 'uuid';
 import pool from '../config/db.js';
 
+// Lấy tất cả 
 export async function getAllCategoriesModel() {
     const [rows] = await pool.query(
-        'SELECT * FROM categories WHERE is_deleted = 0'
+        `SELECT *
+         FROM categories
+         WHERE is_deleted = 0`
     );
     return rows;
 }
 
+// Lấy theo id
 export async function getCategoryByIdModel(id) {
     const [rows] = await pool.query(
-        `SELECT * FROM categories
+        `SELECT *
+         FROM categories
          WHERE category_id = ? AND is_deleted = 0`,
         [id]
     );
     return rows[0] || null;
 }
 
+// Thêm 
 export async function insertCategoryModel(data) {
     const { name } = data;
     const category_id = uuidv7();
@@ -30,36 +36,57 @@ export async function insertCategoryModel(data) {
     return category_id;
 }
 
+// Sửa
 export async function updateCategoryModel(id, data) {
     const { name } = data;
 
     const [result] = await pool.query(
         `UPDATE categories
          SET name = ?
-         WHERE category_id = ?`,
+         WHERE category_id = ? AND is_deleted = 0`,
         [name, id]
     );
 
     return result.affectedRows;
 }
 
-export async function softDeleteCategoryModel(id) {
-    const [result] = await pool.query(
-        `UPDATE categories
-         SET is_deleted = 1
-         WHERE category_id = ?`,
+//  Xoa
+export async function deleteCategoryModel(id) {
+    // ? check ràng buộc
+    const [itemRows] = await pool.query(
+        `SELECT COUNT(*) as count 
+         FROM items
+         WHERE category_id = ? 
+        `,
         [id]
     );
+    const isInItem = itemRows[0].count > 0;
 
-    return result.affectedRows;
-}
+    if (isInItem) {
+        // ? xóa mềm
+        const [rows] = await pool.query(
+            `UPDATE categories
+             SET is_deleted = 1
+             WHERE category_id = ? AND is_deleted = 0`,
+            [id]
+        );
 
-export async function hardDeleteCategoryModel(id) {
-    const [result] = await pool.query(
+
+        return {
+            status: 'soft_deleted',
+            affectedRows: rows.affectedRows
+        }
+    }
+
+    const [rows] = await pool.query(
         `DELETE FROM categories
-         WHERE category_id = ?`,
+         WHERE category_id = ?
+        `,
         [id]
     );
 
-    return result.affectedRows;
-}
+    return {
+        status: 'hard_deleted',
+        affectedRows: rows.affectedRows
+    };
+}
